refactor(cli): split issue/list.rs into list, view, comments, format
No behavior change. Motion only:
- src/cli/issue/view.rs (new) — handle_view + its imports
- src/cli/issue/comments.rs (new) — handle_comments + its imports
- src/cli/issue/format.rs — gains format_comment_date, format_comment_row,
  comment_visibility (+ their 5 inline tests), now shared by view and
  comments via pub(super)
- src/cli/issue/list.rs — reduced to handle_list + list-only helpers
  (build_jql_base_parts, build_filter_clauses, FilterOptions,
  resolve_show_points, extract_unique_status_names) + their 25 tests
- src/cli/issue/mod.rs — register view and comments modules, route
  View/Comments to them

Addresses the CLAUDE.md call-out that list.rs had grown to ~1500 lines
containing three unrelated handlers. File size dropped from 1498 to 1083
lines; cross-handler coupling (handle_view and handle_comments sharing
helpers that lived alongside handle_list) is eliminated.

Note: the spec's estimate that handle_view was ~739 lines was wrong — it
is actually ~268. The real large handler is handle_list (~600 lines of
JQL/filter composition). Extracting list-only helpers is a deferred
follow-up per YAGNI and the spec's Out-of-Scope section.

Validated with Perplexity: pub(super) is the correct visibility for
intra-module sibling access; atomic commit is standard for pure-motion
refactors per rustc / Cargo / ripgrep contributor guidelines.
