+++ src/cli/issue/list.rs
@@ -1,9 +1,7 @@
 use anyhow::Result;
 
-use crate::adf;
 use crate::api::assets::linked::{
-    cmdb_field_ids, enrich_assets, enrich_json_assets, extract_linked_assets,
-    get_or_fetch_cmdb_fields,
+    cmdb_field_ids, enrich_json_assets, extract_linked_assets, get_or_fetch_cmdb_fields,
 };
 use crate::api::client::JiraClient;
 use crate::cli::{IssueCommand, OutputFormat, resolve_effective_limit};
@@ -11,8 +9,6 @@ use crate::config::Config;
 use crate::error::JrError;
 use crate::output;
 use crate::types::assets::LinkedAsset;
-use crate::types::assets::linked::format_linked_assets;
-use crate::types::jira::issue::Comment;
 
 use super::format;
 use super::helpers;
@@ -652,379 +648,6 @@ fn build_filter_clauses(opts: FilterOptions<'_>) -> Vec<String> {
     parts
 }
 
-// ── Comments ─────────────────────────────────────────────────────────
-
-fn format_comment_date(iso: &str, verbose: bool) -> String {
-    use std::sync::atomic::AtomicBool;
-    static LOGGED: AtomicBool = AtomicBool::new(false);
-    match chrono::DateTime::parse_from_rfc3339(iso)
-        .or_else(|_| chrono::DateTime::parse_from_str(iso, "%Y-%m-%dT%H:%M:%S%.3f%z"))
-    {
-        Ok(dt) => dt.format("%Y-%m-%d %H:%M").to_string(),
-        Err(_) => {
-            // Label is "date" (not "comment") because this formatter is also
-            // used by `handle_view` for the Created/Updated timestamp rows.
-            crate::observability::log_parse_failure_once(&LOGGED, "date", iso, verbose);
-            iso.to_string()
-        }
-    }
-}
-
-fn format_comment_row(
-    author_name: Option<&str>,
-    created: Option<&str>,
-    body_text: Option<&str>,
-    verbose: bool,
-) -> Vec<String> {
-    vec![
-        author_name.unwrap_or("(unknown)").to_string(),
-        created
-            .map(|c| format_comment_date(c, verbose))
-            .unwrap_or_else(|| "-".into()),
-        body_text.unwrap_or("(no content)").to_string(),
-    ]
-}
-
-/// Extract the internal/external visibility from a comment's `sd.public.comment` property.
-/// Returns `Some("Internal")` or `Some("External")` if the property exists, `None` otherwise.
-fn comment_visibility(comment: &Comment) -> Option<&'static str> {
-    comment
-        .properties
-        .iter()
-        .find(|p| p.key == "sd.public.comment")
-        .map(|p| {
-            if p.value.get("internal") == Some(&serde_json::Value::Bool(true)) {
-                "Internal"
-            } else {
-                "External"
-            }
-        })
-}
-
-pub(super) async fn handle_comments(
-    key: &str,
-    limit: Option<u32>,
-    output_format: &OutputFormat,
-    client: &JiraClient,
-) -> Result<()> {
-    let comments = client.list_comments(key, limit).await?;
-
-    // Show Visibility column only if any comment has sd.public.comment property
-    let has_visibility = comments.iter().any(|c| comment_visibility(c).is_some());
-
-    match output_format {
-        OutputFormat::Json => {
-            output::print_output(output_format, &[], &[], &comments)?;
-        }
-        OutputFormat::Table => {
-            let verbose = client.verbose();
-            let (headers, rows) = if has_visibility {
-                let rows: Vec<Vec<String>> = comments
-                    .iter()
-                    .map(|c| {
-                        let author = c.author.as_ref().map(|a| a.display_name.as_str());
-                        let created = c.created.as_deref();
-                        let body_text = c.body.as_ref().map(adf::adf_to_text);
-                        let visibility = comment_visibility(c).unwrap_or("External");
-                        let mut row =
-                            format_comment_row(author, created, body_text.as_deref(), verbose);
-                        // Insert Visibility before Body (index 2)
-                        row.insert(2, visibility.to_string());
-                        row
-                    })
-                    .collect();
-                (vec!["Author", "Date", "Visibility", "Body"], rows)
-            } else {
-                let rows: Vec<Vec<String>> = comments
-                    .iter()
-                    .map(|c| {
-                        let author = c.author.as_ref().map(|a| a.display_name.as_str());
-                        let created = c.created.as_deref();
-                        let body_text = c.body.as_ref().map(adf::adf_to_text);
-                        format_comment_row(author, created, body_text.as_deref(), verbose)
-                    })
-                    .collect();
-                (vec!["Author", "Date", "Body"], rows)
-            };
-
-            output::print_output(output_format, &headers, &rows, &comments)?;
-        }
-    }
-
-    Ok(())
-}
-
-// ── View ──────────────────────────────────────────────────────────────
-
-pub(super) async fn handle_view(
-    command: IssueCommand,
-    output_format: &OutputFormat,
-    config: &Config,
-    client: &JiraClient,
-) -> Result<()> {
-    let IssueCommand::View { key } = command else {
-        unreachable!()
-    };
-
-    let sp_field_id = config.global.fields.story_points_field_id.as_deref();
-    let team_field_id: Option<&str> = config.global.fields.team_field_id.as_deref();
-    let cmdb_fields = get_or_fetch_cmdb_fields(client).await.unwrap_or_default();
-    let extra_owned = super::helpers::compose_extra_fields(config, &cmdb_fields);
-    let extra: Vec<&str> = extra_owned.iter().map(String::as_str).collect();
-    let mut issue = client.get_issue(&key, &extra).await?;
-
-    // Extract and enrich assets per-field (shared by both JSON and table paths).
-    // Iterate cmdb_fields directly so we always have (field_id, field_name) together —
-    // avoids any name-based reverse lookups that could break with duplicate field names.
-    let per_field_enriched: Vec<(String, String, Vec<LinkedAsset>)> = if !cmdb_fields.is_empty() {
-        // Extract per-field, keeping both ID and name
-        let mut per_field: Vec<(String, String, Vec<LinkedAsset>)> = Vec::new();
-        for (field_id, field_name) in &cmdb_fields {
-            let assets = extract_linked_assets(&issue.fields.extra, std::slice::from_ref(field_id));
-            if !assets.is_empty() {
-                per_field.push((field_id.clone(), field_name.clone(), assets));
-            }
-        }
-
-        // Collect all assets for batch enrichment
-        let mut all_assets: Vec<LinkedAsset> = per_field
-            .iter()
-            .flat_map(|(_, _, assets)| assets.clone())
-            .collect();
-        enrich_assets(client, &mut all_assets).await;
-
-        // Redistribute enriched assets back
-        let mut enriched = Vec::new();
-        let mut offset = 0;
-        for (field_id, field_name, original_assets) in &per_field {
-            let count = original_assets.len();
-            let assets = all_assets[offset..offset + count].to_vec();
-            offset += count;
-            enriched.push((field_id.clone(), field_name.clone(), assets));
-        }
-        enriched
-    } else {
-        Vec::new()
-    };
-
-    match output_format {
-        OutputFormat::Json => {
-            // Inject enriched data back into JSON before printing
-            if !per_field_enriched.is_empty() {
-                let per_field_by_id: Vec<(String, Vec<LinkedAsset>)> = per_field_enriched
-                    .iter()
-                    .map(|(id, _, assets)| (id.clone(), assets.clone()))
-                    .collect();
-                enrich_json_assets(&mut issue.fields.extra, &per_field_by_id);
-            }
-            println!("{}", output::render_json(&issue)?);
-        }
-        OutputFormat::Table => {
-            let desc_text = issue
-                .fields
-                .description
-                .as_ref()
-                .map(adf::adf_to_text)
-                .unwrap_or_else(|| "(no description)".into());
-
-            let mut rows = vec![
-                vec!["Key".into(), issue.key.clone()],
-                vec!["Summary".into(), issue.fields.summary.clone()],
-                vec![
-                    "Type".into(),
-                    issue
-                        .fields
-                        .issue_type
-                        .as_ref()
-                        .map(|t| t.name.clone())
-                        .unwrap_or_default(),
-                ],
-                vec![
-                    "Status".into(),
-                    issue
-                        .fields
-                        .status
-                        .as_ref()
-                        .map(|s| s.name.clone())
-                        .unwrap_or_default(),
-                ],
-                vec![
-                    "Priority".into(),
-                    issue
-                        .fields
-                        .priority
-                        .as_ref()
-                        .map(|p| p.name.clone())
-                        .unwrap_or_default(),
-                ],
-                vec![
-                    "Assignee".into(),
-                    issue
-                        .fields
-                        .assignee
-                        .as_ref()
-                        .map(|a| a.display_name.clone())
-                        .unwrap_or_else(|| "Unassigned".into()),
-                ],
-                vec![
-                    "Reporter".into(),
-                    issue
-                        .fields
-                        .reporter
-                        .as_ref()
-                        .map(|r| r.display_name.clone())
-                        .unwrap_or_else(|| "(none)".into()),
-                ],
-                vec![
-                    "Created".into(),
-                    issue
-                        .fields
-                        .created
-                        .as_deref()
-                        .map(|c| format_comment_date(c, client.verbose()))
-                        .unwrap_or_else(|| "-".into()),
-                ],
-                vec![
-                    "Updated".into(),
-                    issue
-                        .fields
-                        .updated
-                        .as_deref()
-                        .map(|c| format_comment_date(c, client.verbose()))
-                        .unwrap_or_else(|| "-".into()),
-                ],
-                vec![
-                    "Project".into(),
-                    issue
-                        .fields
-                        .project
-                        .as_ref()
-                        .map(|p| format!("{} ({})", p.name.as_deref().unwrap_or(""), p.key))
-                        .unwrap_or_default(),
-                ],
-                vec![
-                    "Labels".into(),
-                    issue
-                        .fields
-                        .labels
-                        .as_ref()
-                        .filter(|l| !l.is_empty())
-                        .map(|l| l.join(", "))
-                        .unwrap_or_else(|| "(none)".into()),
-                ],
-            ];
-
-            rows.push(vec![
-                "Parent".into(),
-                issue
-                    .fields
-                    .parent
-                    .as_ref()
-                    .map(|p| {
-                        let summary = p
-                            .fields
-                            .as_ref()
-                            .and_then(|f| f.summary.as_deref())
-                            .unwrap_or("");
-                        format!("{} ({})", p.key, summary)
-                    })
-                    .unwrap_or_else(|| "(none)".into()),
-            ]);
-
-            let links_display = issue
-                .fields
-                .issuelinks
-                .as_ref()
-                .filter(|links| !links.is_empty())
-                .map(|links| {
-                    links
-                        .iter()
-                        .map(|link| {
-                            if let Some(ref outward) = link.outward_issue {
-                                let desc = link
-                                    .link_type
-                                    .outward
-                                    .as_deref()
-                                    .unwrap_or(&link.link_type.name);
-                                let summary = outward
-                                    .fields
-                                    .as_ref()
-                                    .and_then(|f| f.summary.as_deref())
-                                    .unwrap_or("");
-                                format!("{} {} ({})", desc, outward.key, summary)
-                            } else if let Some(ref inward) = link.inward_issue {
-                                let desc = link
-                                    .link_type
-                                    .inward
-                                    .as_deref()
-                                    .unwrap_or(&link.link_type.name);
-                                let summary = inward
-                                    .fields
-                                    .as_ref()
-                                    .and_then(|f| f.summary.as_deref())
-                                    .unwrap_or("");
-                                format!("{} {} ({})", desc, inward.key, summary)
-                            } else {
-                                link.link_type.name.clone()
-                            }
-                        })
-                        .collect::<Vec<_>>()
-                        .join("\n")
-                })
-                .unwrap_or_else(|| "(none)".into());
-            rows.push(vec!["Links".into(), links_display]);
-
-            // Per-field asset rows (replaces the old single "Assets" row)
-            for (_, field_name, assets) in &per_field_enriched {
-                let display = format_linked_assets(assets);
-                rows.push(vec![field_name.clone(), display]);
-            }
-
-            if let Some(field_id) = sp_field_id {
-                let points_display = issue
-                    .fields
-                    .story_points(field_id)
-                    .map(format::format_points)
-                    .unwrap_or_else(|| "(none)".into());
-                rows.push(vec!["Points".into(), points_display]);
-            }
-
-            if let Some(field_id) = team_field_id {
-                if let Some(team_uuid) = issue.fields.team_id(field_id, client.verbose()) {
-                    let team_display = match crate::cache::read_team_cache() {
-                        Ok(Some(c)) => c
-                            .teams
-                            .into_iter()
-                            .find(|t| t.id == team_uuid)
-                            .map(|t| t.name)
-                            .unwrap_or_else(|| {
-                                format!(
-                                    "{} (name not cached — run 'jr team list --refresh')",
-                                    team_uuid
-                                )
-                            }),
-                        Ok(None) => format!(
-                            "{} (name not cached — run 'jr team list --refresh')",
-                            team_uuid
-                        ),
-                        Err(e) => {
-                            eprintln!("warning: failed to read team cache: {e}");
-                            format!("{} (team cache unreadable)", team_uuid)
-                        }
-                    };
-                    rows.push(vec!["Team".into(), team_display]);
-                }
-            }
-
-            rows.push(vec!["Description".into(), desc_text]);
-
-            println!("{}", output::render_table(&["Field", "Value"], &rows));
-        }
-    }
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1047,44 +670,6 @@ mod tests {
     fn resolve_show_points_flag_true_config_missing() {
         // Warning emitted to stderr (not captured), but function returns None without error
         assert_eq!(resolve_show_points(true, None), None);
-    }
-
-    #[test]
-    fn format_comment_date_rfc3339() {
-        assert_eq!(
-            format_comment_date("2026-03-20T14:32:00+00:00", false),
-            "2026-03-20 14:32"
-        );
-    }
-
-    #[test]
-    fn format_comment_date_jira_offset_no_colon() {
-        assert_eq!(
-            format_comment_date("2026-03-20T14:32:00.000+0000", false),
-            "2026-03-20 14:32"
-        );
-    }
-
-    #[test]
-    fn format_comment_date_malformed_returns_raw() {
-        assert_eq!(format_comment_date("not-a-date", false), "not-a-date");
-    }
-
-    #[test]
-    fn format_comment_row_missing_author() {
-        let row = format_comment_row(None, Some("2026-03-20T14:32:00+00:00"), None, false);
-        assert_eq!(row[0], "(unknown)");
-    }
-
-    #[test]
-    fn format_comment_row_missing_body() {
-        let row = format_comment_row(
-            Some("Jane Smith"),
-            Some("2026-03-20T14:32:00+00:00"),
-            None,
-            false,
-        );
-        assert_eq!(row[2], "(no content)");
     }
 
     #[test]
